// src/app/theme/shared/models/cliente.ts
export interface Cliente {
    id_empr: string;
    nombre_comercial: string;
    contacto: string;
    id_pais: string;
    id_departamento: string;
    id_localidad: string;
    direccion: string;
    telefonos:string;
    limite_credito: number;
    nit: number;
    nombre_factura: string;
    categoria:string;
}